removed avg time to vbms from table (#8068)
* removed avg time to vbms from second table due to its dependence on the updated_at column which was updated by the upload_remover.rb job and skewed the calculation

* fixed misspelling

* refactored code

* linter change

+++ modules/vba_documents/app/workers/vba_documents/upload_remover.rb
@@ -18,14 +18,13 @@ module VBADocuments
     def perform
       return unless Settings.vba_documents.s3.enabled
 
-      rows = VBADocuments::UploadSubmission.where(REMOVAL_QUERY, EXPIRATION_TIME.ago).order(created_at: :asc).limit(100)
+      rows = VBADocuments::UploadSubmission.where(REMOVAL_QUERY, EXPIRATION_TIME.ago)
 
       rows.find_each do |upload|
         Rails.logger.info('VBADocuments: Cleaning up s3: ' + upload.inspect)
-        next unless store.object(upload.guid).exists?
-
-        store.delete(upload.guid)
-        upload.update(s3_deleted: true)
+        store.delete(upload.guid) if store.object(upload.guid).exists?
+        upload.s3_deleted = true
+        upload.save!(touch: false)
       end
     end
